feat(rename): rename app to AudioTask Reminder and package to com.apps.audiotaskreminder
